Disallow `==` in contrasts. Instead handle `-` correctly
This makes lemur consistent with glmGamPoi if the contrast is specified
with `cond()`. Because lemur works on the prediction level (unlike GLMs)
the traditional way to just specify the contrast vector doesn't make
sense and accordingly triggers an error now.

Also remove the option to specify an alignment_contrast in `test_de()`.
This seemed overly complicated and I want to see an application before
allowing it again.