Fix bug in subsetting logic.
Previously, I only tracked if a gene was supposed to be still present
after subsetting. However, if the order of genes changed `predict`
did not reflect that.

I modified the `row_mask` to account for this. Originally, it was a
boolean vector indicating if a gene was present. Now it is an integer
vector where each value points to row in the coefficient matrix.

This bug fixes affects predictions (including test_de) calls on
subsetted fit objects.